requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
distro==1.8.0
Django==4.2.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jsonschema==3.2.0
packaging==23.1
paramiko==3.2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
typing_extensions==4.6.3
urllib3==2.0.3
websocket-client==0.59.0
